Refactored StudentGroups component
