+++ src/authenticated-app.tsx
@@ -4,14 +4,14 @@ import { Switch, Route } from 'react-router-dom'
 
 import StudentGroups from 'screens/StudentGroups'
 import StudentGroup from 'screens/StudentGroup'
-import ManageStudents from 'screens/ManageStudents'
+import ShowAllStudents from 'screens/ShowAllStudents'
 
 const AuthenticatedApp: React.FC = () => {
   return (
     <Switch>
       <Route path="/" exact component={StudentGroups} />
       <Route path="/student-group/:groupId" component={StudentGroup} />
-      <Route path="/manage-student-group/:groupId" component={ManageStudents} />
+      <Route path="/show-all-students/:groupId" component={ShowAllStudents} />
     </Switch>
   )
 }
